requirements.txt
streamlit
sounddevice
soundfile
python-dotenv
groq
RPi.GPIO
edge-tts